requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
google-generativeai==0.3.0
pydantic==2.5.0
python-dotenv==1.0.0
gunicorn